+++ textsearch/python/csrc/utils.cc
@@ -31,7 +31,7 @@ Args:
 >>> import numpy as np
 >>> keep = np.array([0, 0, 1, 0, 1, 0, 1, 1], dtype=bool)
 >>> get_new2old(keep)
-array([2, 4, 6, 7], dtype=uint64)
+array([2, 4, 6, 7], dtype=uint32)
 
 )doc";
 
@@ -58,11 +58,11 @@ static void PybindRowIdsToRowSplits(py::module &m) {
 static void PybindGetNew2Old(py::module &m) {
   m.def(
       "get_new2old",
-      [](py::array_t<bool> keep) -> py::array_t<uint64_t> {
+      [](py::array_t<bool> keep) -> py::array_t<uint32_t> {
         py::buffer_info keep_buf = keep.request();
         size_t num_old_elems = keep_buf.size;
         const bool *p_keep = static_cast<const bool *>(keep_buf.ptr);
-        std::vector<uint64_t> new2old;
+        std::vector<uint32_t> new2old;
         GetNew2Old(p_keep, num_old_elems, &new2old);
         return py::array(new2old.size(), new2old.data());
       },
